python: Factor out python2 specific code from trace-python.c
In order to support both python2 and python3, this patch adds a new
"trace-python2.c" file, which contains only python2 specific code.

"trace-python.c" only contains common code for both python2 and python3.
